refactor: - improved OCP and DIP
