+++ src/main/java/edu/kis/powp/jobs2d/command/parser/ManualCsvParser.java
@@ -1,0 +1,7 @@
+package edu.kis.powp.jobs2d.command.parser;
+
+public class ManualCsvParser implements ManualParser {
+
+
+
+}
